Some environment cleanup, and a new class. Remove the ability to delete indices in the frontend or admin API.

+++ bin/elasticsearch.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
 
+# Script to automatically handle an Elasticsearch install or upgrade.
+# Downloads and installs the update while preserving primary index data.
+# Non-primary indices will be repopulated when OL starts again.
+
 prog=$(basename "$0")
 indices=("notifications" "apilog" "bills" "agendas" "calendars" "committees" "transcripts" "hearings" "laws" "members")
 usage() {
-  echo "Usage: $prog [--delete-indices] [--new-version es_version_number]" >&2
+  echo "Usage: $prog [--install] [--delete-indices] [--new-version es_version_number]" >&2
 }
+
 check_exit_code() {
   if [ $? != 0 ]; then
     exit 1;
   fi
 }
-#TODO: check for superuser privileges
+
 es_reindex() {
   echo "Reindexing documents from $1 to $2..."
   curl -HContent-Type:application/json -XPOST localhost:9200/_reindex?pretty -d'{
@@ -21,24 +26,29 @@ es_reindex() {
       "index": "'"$2"'"
     }
   }'
-  check_exit_code
   echo "Done reindexing."
 }
 
 es_delete() {
-  # TODO: check doesn't work
-  if curl -I "localhost:9200/$1"; then
+  http_code=$(curl -s -o /dev/null -w "%{http_code}" "localhost:9200/$1")
+  if [ "$http_code" = 200 ]; then
     curl -X DELETE "localhost:9200/$1"
-    check_exit_code
     echo "Deleted $1 index."
   fi
 }
 
-delete=0
-new_version=0
+if [[ $EUID -ne 0 ]]
+  then echo "Please run as root."
+  exit 1
+fi
+
+install=""
+delete=""
+new_version=""
 
 while [ $# -gt 0 ]; do
   case "$1" in
+    --install) install=1 ;;
     --delete-indices) delete=1 ;;
     --new-version) new_version="$2"; shift; ;;
     --help) usage; exit 0 ;;
@@ -47,7 +57,12 @@ while [ $# -gt 0 ]; do
   shift
 done
 
-if [ "$new_version" == 0 ]; then
+if [ "$install" ]; then
+    echo "Will install Elasticsearch and ignore other options."
+    delete=""
+    # Grabs the ES version from the pom.
+    new_version=$(grep '<elasticsearch.version>' ../pom.xml | grep -o '[0-9][0-9.]*')
+elif [ ! "$new_version" ]; then
     usage
     exit 1
 fi
@@ -55,7 +70,7 @@ fi
 wget https://artifacts.elastic.co/downloads/elasticsearch/elasticsearch-"${new_version}"-amd64.deb >&1
 check_exit_code
 
-if [ "$delete" != 0 ]; then
+if [ "$delete" ]; then
     for primary_store in "${indices[0]}" "${indices[1]}"; do
       es_reindex "$primary_store" "$primary_store"-old
     done
@@ -64,22 +79,26 @@ if [ "$delete" != 0 ]; then
     done
 fi
 
-# A flush should speed things up
-curl -X POST "localhost:9200/_flush"
-systemctl stop elasticsearch.service
+if [ ! "$install" ]; then
+  # A flush should speed things up.
+  curl -X POST "localhost:9200/_flush"
+  systemctl stop elasticsearch.service
+fi
 dpkg -i elasticsearch-"${new_version}"-amd64.deb >&1
 check_exit_code
 systemctl daemon-reload
+if [ "$install" ]; then
+    systemctl enable elasticsearch.service
+fi
 systemctl start elasticsearch.service
 rm elasticsearch-"${new_version}"-amd64.deb
 
-read -rp "Restart OpenLegislation and wait for index creation to finish. Press enter to continue."
-
-# Primary stores must be restored, and the backups deleted
-if [ "$delete" != 0 ]; then
+# Primary stores must be restored, and the backups deleted.
+if [ "$delete" ]; then
+  read -rp "Restart OpenLegislation and wait for index creation to finish. Then, press enter to continue."
   for primary_store in "${indices[0]}" "${indices[1]}"; do
     es_reindex "$primary_store"-old "$primary_store"
     es_delete "$primary_store"-old
   done
+  echo "Primary store indices have been restored."
 fi
-echo "Primary store indices have been restored."
